requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
PyNaCl>=1.5.0
aiofiles>=23.2.1
pydub>=0.25.1
